requirements.txt
Flask==2.3.3
PyMuPDF==1.23.8
python-docx==0.8.11
diff-match-patch==20230430
Werkzeug==2.3.7
